it83xx/adc: Use 64-bit int for voltage calculation
For some voltage dividers, factor_mul is large enough that the
intermediate value can overflow, so use a 64-bit int.

BRANCH=None
BUG=b:271941405
TEST=VBUS ADC calculation doesn't overflow on dibbi

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4506010
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>

+++ chip/it83xx/adc.c
@@ -167,7 +167,7 @@ int adc_read_channel(enum adc_channel ch)
 		else
 			IT83XX_ADC_ADCDVSTS2 = (1 << (adc_ch - CHIP_ADC_CH13));
 
-		mv = adc_raw_data * adc_channels[ch].factor_mul /
+		mv = (uint64_t)adc_raw_data * adc_channels[ch].factor_mul /
 			     adc_channels[ch].factor_div +
 		     adc_channels[ch].shift;
 		valid = 1;
